Decouple disposing LanguageWorkerChannel and WorkerPorcess
