+++ src/WebJobs.Script/Rpc/ILanguageWorkerChannel.cs
@@ -20,5 +20,7 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
         void SendFunctionEnvironmentReloadRequest();
 
         void StartWorkerProcess();
+
+        void ShutdownWorkerProcess();
     }
 }
