+++ src/WebJobs.Script/Rpc/LanguageWorkerChannel.cs
@@ -10,7 +10,6 @@ using System.IO;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Threading.Tasks.Dataflow;
-using Microsoft.Azure.WebJobs.Logging;
 using Microsoft.Azure.WebJobs.Script.Description;
 using Microsoft.Azure.WebJobs.Script.Diagnostics;
 using Microsoft.Azure.WebJobs.Script.Eventing;
@@ -29,20 +28,18 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
         private readonly TimeSpan workerInitTimeout = TimeSpan.FromSeconds(30);
         private readonly string _rootScriptPath;
         private readonly IScriptEventManager _eventManager;
-        private readonly IWorkerProcessFactory _processFactory;
-        private readonly IProcessRegistry _processRegistry;
         private readonly WorkerConfig _workerConfig;
         private readonly ILogger _workerChannelLogger;
         private readonly ILanguageWorkerConsoleLogSource _consoleLogSource;
+        private readonly IWorkerProcessFactory _processFactory;
+        private readonly IProcessRegistry _processRegistry;
 
         private bool _disposed;
-        private bool _disposing;
         private bool _isWebHostChannel;
         private IObservable<FunctionRegistrationContext> _functionRegistrations;
         private WorkerInitResponse _initMessage;
         private string _workerId;
-        private Process _process;
-        private Queue<string> _processStdErrDataQueue = new Queue<string>(3);
+        private LanguageWorkerProcess _languageWorkerProcess;
         private IDictionary<string, BufferBlock<ScriptInvocationContext>> _functionInputBuffers = new Dictionary<string, BufferBlock<ScriptInvocationContext>>();
         private IDictionary<string, Exception> _functionLoadErrors = new Dictionary<string, Exception>();
         private ConcurrentDictionary<string, ScriptInvocationContext> _executingInvocations = new ConcurrentDictionary<string, ScriptInvocationContext>();
@@ -77,13 +74,13 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
             _functionRegistrations = functionRegistrations;
             _rootScriptPath = rootScriptPath;
             _eventManager = eventManager;
-            _processFactory = processFactory;
-            _processRegistry = processRegistry;
             _workerConfig = workerConfig;
             _serverUri = serverUri;
             _isWebHostChannel = isWebHostChannel;
             _workerChannelLogger = loggerFactory.CreateLogger($"Worker.{workerConfig.Language}.{_workerId}");
             _consoleLogSource = consoleLogSource;
+            _processRegistry = processRegistry;
+            _processFactory = processFactory;
 
             _inboundWorkerEvents = _eventManager.OfType<InboundEvent>()
                 .Where(msg => msg.WorkerId == _workerId);
@@ -110,126 +107,9 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
 
         public WorkerConfig Config => _workerConfig;
 
-        internal Queue<string> ProcessStdErrDataQueue => _processStdErrDataQueue;
-
-        internal Process WorkerProcess => _process;
+        internal Process WorkerProcess => _languageWorkerProcess.WorkerProcess;
 
         public bool IsWebhostChannel => _isWebHostChannel;
-
-        internal void StartProcess()
-        {
-            try
-            {
-                _process.ErrorDataReceived += (sender, e) => OnErrorDataReceived(sender, e);
-                _process.OutputDataReceived += (sender, e) => OnOutputDataReceived(sender, e);
-                _process.Exited += (sender, e) => OnProcessExited(sender, e);
-                _process.EnableRaisingEvents = true;
-
-                _workerChannelLogger?.LogInformation($"Starting language worker process:{_process.StartInfo.FileName} {_process.StartInfo.Arguments}");
-                _process.Start();
-                _workerChannelLogger?.LogInformation($"{_process.StartInfo.FileName} process with Id={_process.Id} started");
-
-                _process.BeginErrorReadLine();
-                _process.BeginOutputReadLine();
-            }
-            catch (Exception ex)
-            {
-                throw new HostInitializationException($"Failed to start Language Worker Channel for language :{_workerConfig.Language}", ex);
-            }
-        }
-
-        private void OnOutputDataReceived(object sender, DataReceivedEventArgs e)
-        {
-            if (e.Data != null)
-            {
-                string msg = e.Data;
-                if (LanguageWorkerChannelUtilities.IsLanguageWorkerConsoleLog(msg))
-                {
-                    msg = LanguageWorkerChannelUtilities.RemoveLogPrefix(msg);
-                    _workerChannelLogger?.LogInformation(msg);
-                }
-                else
-                {
-                    _consoleLogSource?.Log(msg);
-                }
-            }
-        }
-
-        private void OnProcessExited(object sender, EventArgs e)
-        {
-            if (_disposing)
-            {
-                // No action needed
-                return;
-            }
-            string exceptionMessage = string.Join(",", _processStdErrDataQueue.Where(s => !string.IsNullOrEmpty(s)));
-            try
-            {
-                if (_process.ExitCode != 0)
-                {
-                    var processExitEx = new LanguageWorkerProcessExitException($"{_process.StartInfo.FileName} exited with code {_process.ExitCode}\n {exceptionMessage}");
-                    processExitEx.ExitCode = _process.ExitCode;
-                    HandleWorkerError(processExitEx);
-                }
-                else
-                {
-                    _process.WaitForExit();
-                    _process.Close();
-                }
-            }
-            catch (Exception)
-            {
-                // ignore process is already disposed
-            }
-        }
-
-        private void OnErrorDataReceived(object sender, DataReceivedEventArgs e)
-        {
-            // TODO: per language stdout/err parser?
-            if (e.Data != null)
-            {
-                string msg = e.Data;
-                if (msg.IndexOf("warn", StringComparison.OrdinalIgnoreCase) > -1)
-                {
-                    if (LanguageWorkerChannelUtilities.IsLanguageWorkerConsoleLog(msg))
-                    {
-                        msg = LanguageWorkerChannelUtilities.RemoveLogPrefix(msg);
-                        _workerChannelLogger?.LogWarning(msg);
-                    }
-                    else
-                    {
-                        _consoleLogSource?.Log(msg);
-                    }
-                }
-                else if ((msg.IndexOf("error", StringComparison.OrdinalIgnoreCase) > -1) ||
-                          (msg.IndexOf("fail", StringComparison.OrdinalIgnoreCase) > -1) ||
-                          (msg.IndexOf("severe", StringComparison.OrdinalIgnoreCase) > -1))
-                {
-                    if (LanguageWorkerChannelUtilities.IsLanguageWorkerConsoleLog(msg))
-                    {
-                        msg = LanguageWorkerChannelUtilities.RemoveLogPrefix(msg);
-                        _workerChannelLogger?.LogError(msg);
-                    }
-                    else
-                    {
-                        _consoleLogSource?.Log(msg);
-                    }
-                    _processStdErrDataQueue = LanguageWorkerChannelUtilities.AddStdErrMessage(_processStdErrDataQueue, Sanitizer.Sanitize(msg));
-                }
-                else
-                {
-                    if (LanguageWorkerChannelUtilities.IsLanguageWorkerConsoleLog(msg))
-                    {
-                        msg = LanguageWorkerChannelUtilities.RemoveLogPrefix(msg);
-                        _workerChannelLogger?.LogInformation(msg);
-                    }
-                    else
-                    {
-                        _consoleLogSource?.Log(msg);
-                    }
-                }
-            }
-        }
 
         public void StartWorkerProcess()
         {
@@ -248,9 +128,13 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
                 ServerUri = _serverUri,
             };
 
-            _process = _processFactory.CreateWorkerProcess(workerContext);
-            StartProcess();
-            _processRegistry?.Register(_process);
+            _languageWorkerProcess = new LanguageWorkerProcess(_workerConfig.Language, _workerId, workerContext, _eventManager, _processFactory, _processRegistry, _workerChannelLogger, _consoleLogSource);
+            _languageWorkerProcess.StartProcess();
+        }
+
+        public void ShutdownWorkerProcess()
+        {
+            _languageWorkerProcess?.ShutdownProcess();
         }
 
         // send capabilities to worker, wait for WorkerInitResponse
@@ -466,16 +350,6 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
 
         internal void HandleWorkerError(Exception exc)
         {
-            LanguageWorkerProcessExitException langExc = exc as LanguageWorkerProcessExitException;
-            // The subscriber of WorkerErrorEvent is expected to Dispose() the errored channel
-            if (langExc != null && langExc.ExitCode == -1)
-            {
-                _workerChannelLogger.LogDebug(exc, $"Language Worker Process exited.", _process.StartInfo.FileName);
-            }
-            else
-            {
-                _workerChannelLogger.LogError(exc, $"Language Worker Process exited.", _process.StartInfo.FileName);
-            }
             _eventManager.Publish(new WorkerErrorEvent(_workerConfig.Language, Id, exc));
         }
 
@@ -499,24 +373,6 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
                         link.Dispose();
                     }
 
-                    // best effort process disposal
-                    try
-                    {
-                        if (_process != null)
-                        {
-                            if (!_process.HasExited)
-                            {
-                                _process.Kill();
-                                _process.WaitForExit();
-                            }
-                            _process.Dispose();
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        _workerChannelLogger.LogError(e, "LanguageWorkerChannel Dispose failure");
-                    }
-
                     foreach (var sub in _eventSubscriptions)
                     {
                         sub.Dispose();
@@ -528,7 +384,6 @@ namespace Microsoft.Azure.WebJobs.Script.Rpc
 
         public void Dispose()
         {
-            _disposing = true;
             Dispose(true);
         }
     }
